Separate commands and queries
