+++ contracts/external/cw-abc/src/lib.rs
@@ -4,5 +4,7 @@ mod error;
 pub mod msg;
 pub mod state;
 pub mod abc;
+pub(crate) mod commands;
+mod queries;
 
 pub use crate::error::ContractError;
